chore: Drops in-memory-backed dataset, moves storage to a JSON storage-pointer
